add basic linear mm allocator

+++ code/include/mm/mm.h
@@ -6,8 +6,8 @@
 #include <mm/layout.h>
 #include <mm/page.h>
 
-typedef uint64_t virt_addr_t;
-typedef uint64_t phys_addr_t;
-typedef uint64_t dma_addr_t;
+typedef size_t virt_addr_t;
+typedef size_t phys_addr_t;
+typedef size_t dma_addr_t;
 
 #endif
